+++ StripePaymentSheet/StripePaymentSheet/Source/Helpers/Images.swift
@@ -43,6 +43,7 @@ enum Image: String, CaseIterable, ImageMaker {
     case pm_type_oxxo = "icon-pm-oxxo"
     case pm_type_konbini = "icon-pm-konbini"
     case pm_type_boleto = "icon-pm-boleto"
+    case pm_type_swish = "icon-pm-swish"
 
     // Icons/symbols
     case icon_checkmark = "icon_checkmark"