+++ StripePayments/StripePayments/Source/Categories/Enums+CustomStringConvertible.swift
@@ -515,6 +515,8 @@ extension STPPaymentMethodType: CustomStringConvertible {
             return "weChatPay"
         case .cashApp:
             return "cashApp"
+        case .swish:
+            return "swish"
         case .paynow, .zip, .revolutPay, .mobilePay, .amazonPay, .alma, .konbini, .promptPay:
             // `description` is the value used when this type is converted to a string for debugging purposes, just use the display name.
             return displayName