+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * [Fixed] Fixed an issue with selecting from lists on macOS Catalyst. Note that only macOS 11 or later is supported: We do not recommend releasing a Catalyst app targeting macOS 10.15.
 * [Fixed] Fixed an issue with scanning card expiration dates.
 * [Fixed] Fixed an issue where billing address collection configuration was not passed to Apple Pay.
+* [Added] Support for Swish with PaymentIntents.
 
 ### Basic Integration
 * [Fixed] Fixed an issue with scanning card expiration dates.
@@ -10,6 +11,7 @@
 ### Payments
 * [Fixed] Fixed an issue where amounts in Serbian Dinar were displayed incorrectly.
 * [Fixed] Fixed an issue where the SDK could hang in macOS Catalyst.
+* [Added] Support for Swish bindings.
 
 ## 23.16.0 2023-09-18
 ### Payments